rename: #59 client 디렉토리 경로 변경
